move VmwareRestModule two level below
Move `ansible_collections.ansible.vmware_rest.plugins.module_utils.vmware_httpapi.VmwareRestModule.VmwareRestModule` in
`ansible_collections.ansible.vmware_rest.plugins.module_utils.vmware_httpapi.VmwareRestModule` to reduce the path
length.
