+++ tests/units/modules/cloud/vmware_httpapi/test_vmware_appliance_access_info.py
@@ -7,6 +7,7 @@ import pytest
 from .conftest import enable_vcr
 
 import six
+
 if six.PY3:
     from unittest.mock import ANY as ANY
 else:
